Rename suite-core runtime-role filenames to active logic role
